requirementsSelfBot.txt
discord.py-self
Flask
selenium
undetected-chromedriver
setuptools
plyvel